requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
aniso8601==9.0.1
anyio==3.7.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.4
django-admin-autocomplete-filter==0.7.1
django-admin-filters==0.1.3
django-admin-list-filter-dropdown==1.0.3
django-adminfilters==1.9.0
django-appconf==1.0.5
django-author==1.2.0
django-birthday==0.1.4
django-bookmark-base==2020.7.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-dynamic-raw-id==3.0
django-environ==0.9.0
django-filter==22.1
django-graphql-jwt==0.3.4
django-import-export==3.0.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-nextjs==2.2.2
django-phone-field==1.8.1
django-select2==8.1.1
django-simple-history==3.3.0
django-tables2==2.4.1
django-taggit==3.1.0
django-templated-mail==1.1.1
django-tinymce==3.5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
drf-yasg==1.21.4
et-xmlfile==1.1.0
frozenlist==1.3.3
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
libcst==1.0.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.1
mdurl==0.1.2
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.3.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.16.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-barcode==0.14.0
python-dateutil==2.8.2
python-multipart==0.0.6
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.6
PyYAML==6.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.5.2
ruamel.yaml==0.17.21
setuptools-git==1.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.4
sqlparse==0.4.3
starlette==0.31.0
strawberry-graphql==0.199.3
strawberry-graphql-django==0.16.0
tablib==3.2.1
text-unidecode==1.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.1
whitenoise==6.4.0
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
